requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
mysqlclient==2.2.7
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
